# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "enum": [
          "bad_request",
          "unauthenticated",
          "token_invalid",
          "account_deleted",
          "membership_required",
          "role_required",
          "limit_reached",
          "not_found",
          "method_not_allowed",
          "invitation_invalid",
          "conflict",
          "sole_owner",
          "invitation_expired",
          "archive_unavailable",
          "validation_failed",
          "unsupported_operation",
          "rate_limited",
          "internal",
          "storage_unavailable"
        ],
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "Feature": {
        "properties": {
          "geometry": {
            "discriminator": {
              "mapping": {
                "LineString": "#/components/schemas/LineStringGeometry",
                "MultiLineString": "#/components/schemas/MultiLineStringGeometry",
                "MultiPoint": "#/components/schemas/MultiPointGeometry",
                "MultiPolygon": "#/components/schemas/MultiPolygonGeometry",
                "Point": "#/components/schemas/PointGeometry",
                "Polygon": "#/components/schemas/PolygonGeometry"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PointGeometry"
              },
              {
                "$ref": "#/components/schemas/MultiPointGeometry"
              },
              {
                "$ref": "#/components/schemas/LineStringGeometry"
              },
              {
                "$ref": "#/components/schemas/MultiLineStringGeometry"
              },
              {
                "$ref": "#/components/schemas/PolygonGeometry"
              },
              {
                "$ref": "#/components/schemas/MultiPolygonGeometry"
              }
            ],
            "title": "Geometry"
          },
          "properties": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Properties",
            "type": "object"
          },
          "type": {
            "const": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "geometry"
        ],
        "title": "Feature",
        "type": "object"
      },
      "FeatureCollection": {
        "properties": {
          "crs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NamedCrs"
              },
              {
                "type": "null"
              }
            ]
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "title": "Features",
            "type": "array"
          },
          "type": {
            "const": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "features"
        ],
        "title": "FeatureCollection",
        "type": "object"
      },
...
      "JsonValue": {},
      "LineStringGeometry": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "LineString",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "coordinates"
        ],
        "title": "LineStringGeometry",
        "type": "object"
      },
...
      "MultiLineStringGeometry": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "MultiLineString",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "coordinates"
        ],
        "title": "MultiLineStringGeometry",
        "type": "object"
      },
      "MultiPointGeometry": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "MultiPoint",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "coordinates"
        ],
        "title": "MultiPointGeometry",
        "type": "object"
      },
      "MultiPolygonGeometry": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "MultiPolygon",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "coordinates"
        ],
        "title": "MultiPolygonGeometry",
        "type": "object"
      },
      "NamedCrs": {
        "properties": {
          "properties": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Properties",
            "type": "object"
          }
        },
        "title": "NamedCrs",
        "type": "object"
      },
...
      "PackageDetail": {
        "properties": {
          "archive_bytes": {
            "title": "Archive Bytes",
            "type": "integer"
          },
          "archive_sha256": {
            "title": "Archive Sha256",
            "type": "string"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/PreparationCheck"
            },
            "title": "Checks",
            "type": "array"
          },
          "form_version": {
            "title": "Form Version",
            "type": "string"
          },
          "manifest": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "package_id": {
            "format": "uuid",
            "title": "Package Id",
            "type": "string"
          },
          "prepared_at": {
            "format": "date-time",
            "title": "Prepared At",
            "type": "string"
          },
          "site_code": {
            "title": "Site Code",
            "type": "string"
          },
          "state": {
            "enum": [
              "ready",
              "blocked"
            ],
            "title": "State",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "package_id",
          "site_code",
          "form_version",
          "version",
          "state",
          "archive_bytes",
          "archive_sha256",
          "prepared_at",
          "manifest",
          "checks"
        ],
        "title": "PackageDetail",
        "type": "object"
      },
      "PackageSubmission": {
        "additionalProperties": false,
        "properties": {
          "form_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Form Version",
            "type": "string"
          },
          "layers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FeatureCollection"
            },
            "propertyNames": {
              "enum": [
                "ground",
                "paths",
                "trees",
                "zones"
              ]
            },
            "title": "Layers",
            "type": "object"
          },
          "project_file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectFile"
              },
              {
                "type": "null"
              }
            ]
          },
          "site_code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Site Code",
            "type": "string"
          }
        },
        "required": [
          "site_code",
          "form_version",
          "layers"
        ],
        "title": "PackageSubmission",
        "type": "object"
      },
      "PackageSummary": {
        "description": "A prepared package as a list shows it. The archive itself is fetched separately.",
        "properties": {
          "archive_bytes": {
            "title": "Archive Bytes",
            "type": "integer"
          },
          "archive_sha256": {
            "title": "Archive Sha256",
            "type": "string"
          },
          "form_version": {
            "title": "Form Version",
            "type": "string"
          },
          "package_id": {
            "format": "uuid",
            "title": "Package Id",
            "type": "string"
          },
          "prepared_at": {
            "format": "date-time",
            "title": "Prepared At",
            "type": "string"
          },
          "site_code": {
            "title": "Site Code",
            "type": "string"
          },
          "state": {
            "enum": [
              "ready",
              "blocked"
            ],
            "title": "State",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "package_id",
          "site_code",
          "form_version",
          "version",
          "state",
          "archive_bytes",
          "archive_sha256",
          "prepared_at"
        ],
        "title": "PackageSummary",
        "type": "object"
      },
      "PointGeometry": {
        "properties": {
          "coordinates": {
            "items": {
              "type": "number"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "Point",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "coordinates"
        ],
        "title": "PointGeometry",
        "type": "object"
      },
      "PolygonGeometry": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "Polygon",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "coordinates"
        ],
        "title": "PolygonGeometry",
        "type": "object"
      },
      "PreparationCheck": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "state": {
            "enum": [
              "passed",
              "warning",
              "blocked",
              "skipped"
            ],
            "title": "State",
            "type": "string"
          },
          "step": {
            "enum": [
              "source-project",
              "layer-sources",
              "coordinate-reference",
              "imagery-licence",
              "archive"
            ],
            "title": "Step",
            "type": "string"
          }
        },
        "required": [
          "step",
          "state",
          "detail"
        ],
        "title": "PreparationCheck",
        "type": "object"
      },
...
      "ProjectFile": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "contentMediaType": "application/octet-stream",
            "format": "base64",
            "title": "Content",
            "type": "string"
          },
          "file_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "File Name",
            "type": "string"
          }
        },
        "required": [
          "file_name",
          "content"
        ],
        "title": "ProjectFile",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FieldMaps API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects/{project_id}/packages": {
      "get": {
        "operationId": "listPackages",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PackageSummary"
                  },
                  "title": "Response Listpackages",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Packages"
      },
      "post": {
        "description": "Prepare a site package from a QGIS export, and record what every check found.",
        "operationId": "preparePackage",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Prepare"
      }
    },
    "/v1/projects/{project_id}/packages/{package_id}": {
      "get": {
        "operationId": "getPackage",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Package"
      }
    },
    "/v1/projects/{project_id}/packages/{package_id}/archive": {
      "get": {
        "description": "Hand over the package itself, for a device about to carry it into the field.",
        "operationId": "downloadPackageArchive",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Prepared package archive",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Quoted SHA-256 digest",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive"
      }
    }
  }
}